Add traits for database and storage adapters
To more easily test combinations of database and storage adapters in the
Behat test suites we can use traits for the different adapters. Each
adapter should be responsible of setup / teardown functionality related
to the different adapters. Each suite could then simply use two traits
in combination to test them.
